func on reducers are all exported out/ cleaning out redux store
