+++ src/types/Book/Book.tsx
@@ -1,0 +1,8 @@
+import Field from "../FieldTypes/FieldTypes";
+export interface Book {
+  id: number;
+  name: Field;
+  author: Field;
+  gender: Field;
+  img: Field;
+}
